SSR refactor, security hardening, and CI/CD improvements
SSR refactor, security hardening, and CI/CD improvements

- modularize the frontend SSR runtime into focused server components
- introduce a shared transport layer for frontend API services
- harden SSR request handling and expand server-side coverage
- resolve frontend audit issues and improve Safety-based security checks
- pin backend dependencies to improve build and vulnerability scan reproducibility
- fix CI/CD validation, status reporting, and production nginx image verification
- tighten `dev -> main` release checks and production image build validation
