+++ .github/workflows/ci-cd.yml
@@ -21,7 +21,7 @@ jobs:
   version-check:
     name: Version bump check (dev → main)
     runs-on: ubuntu-latest
-    if: github.event_name == 'pull_request' && github.base_ref == 'main'
+    if: github.event_name == 'pull_request' && github.base_ref == 'main' && github.head_ref == 'dev'
     steps:
       - name: Checkout PR HEAD
         uses: actions/checkout@v4
@@ -32,10 +32,10 @@ jobs:
         run: |
           if [[ -f VERSION ]]; then
             VER=$(tr -d ' \t\r\n' < VERSION)
-            echo "VERSION: $VER"
-            echo "Release tag would be: v$VER"
+            echo "✅ PR introduces VERSION: $VER"
+            echo "🏷️  Release tag would be: v$VER"
           else
-            echo "VERSION file not found"
+            echo "⚠️ VERSION file not found"
           fi
 
       - name: Validate VERSION file and bump
@@ -48,46 +48,59 @@ jobs:
 
           VERSION_FILE="VERSION"
 
+          echo "🔎 Version check (dev → main)"
+          echo "📌 Base SHA: $BASE_SHA"
+          echo "📌 Head SHA: $HEAD_SHA"
+
+          # PR/DEV must have VERSION
           if ! git cat-file -e "${HEAD_SHA}:${VERSION_FILE}" 2>/dev/null; then
-            echo "ERROR: ${VERSION_FILE} is missing in PR branch."
+            echo "❌ ERROR: ${VERSION_FILE} is missing in PR branch (dev)."
             exit 1
           fi
 
           NEW="$(git show "${HEAD_SHA}:${VERSION_FILE}" | tr -d ' \t\r\n')"
           if ! [[ "$NEW" =~ ^[0-9]+\.[0-9]+\.[0-9]+$ ]]; then
-            echo "ERROR: VERSION must be MAJOR.MINOR.PATCH. Got: '$NEW'"
+            echo "❌ ERROR: VERSION must be MAJOR.MINOR.PATCH (e.g. 1.0.2). Got: '$NEW'"
             exit 1
           fi
 
+          # If main/base doesn't have VERSION yet -> bootstrap pass
           if ! git cat-file -e "${BASE_SHA}:${VERSION_FILE}" 2>/dev/null; then
-            echo "Bootstrap: ${VERSION_FILE} does not exist on base branch."
-            echo "VERSION introduced: $NEW"
+            echo "⚠️  Bootstrap: ${VERSION_FILE} does not exist on base branch (main)."
+            echo "✅ PR introduces VERSION: $NEW"
+            echo "🏷️  Release tag would be: v$NEW"
             exit 0
           fi
 
           OLD="$(git show "${BASE_SHA}:${VERSION_FILE}" | tr -d ' \t\r\n')"
           if ! [[ "$OLD" =~ ^[0-9]+\.[0-9]+\.[0-9]+$ ]]; then
-            echo "ERROR: Base VERSION is invalid. Got: '$OLD'"
+            echo "❌ ERROR: Base VERSION is invalid on main. Got: '$OLD'"
             exit 1
           fi
 
+          echo "🔢 Base version: $OLD"
+          echo "🔢 PR version:   $NEW"
+
           if [[ "$OLD" == "$NEW" ]]; then
-            echo "ERROR: VERSION was not changed in this PR."
+            echo "❌ ERROR: VERSION was not changed in this PR (still $NEW)."
             exit 1
           fi
 
           IFS='.' read -r oldM oldm oldp <<< "$OLD"
           IFS='.' read -r newM newm newp <<< "$NEW"
 
+          # Require strictly greater
           if (( newM < oldM )) || \
             (( newM == oldM && newm < oldm )) || \
             (( newM == oldM && newm == oldm && newp <= oldp )); then
-            echo "ERROR: VERSION must increase. Base=$OLD PR=$NEW"
+            echo "❌ ERROR: VERSION must increase. Base=$OLD PR=$NEW"
             exit 1
           fi
 
-          echo "VERSION bump is valid: $OLD -> $NEW"
+          echo "✅ VERSION bump is valid: $OLD → $NEW"
+          echo "🏷️  Release tag would be: v$NEW"
 
+  # Frontend: Lint, Type Check, and Test
   frontend:
     name: Frontend Quality & Tests
     runs-on: ubuntu-latest
@@ -132,9 +145,11 @@ jobs:
           flags: frontend
           token: ${{ secrets.CODECOV_TOKEN }}
 
+  # Backend: Lint, Type Check, and Test
   backend:
     name: Backend Quality & Tests
     runs-on: ubuntu-latest
+
     services:
       postgres:
         image: postgres:18
@@ -157,6 +172,8 @@ jobs:
       DEFAULT_FROM_EMAIL: test@example.com
       CONTACT_EMAIL: test@example.com
       SESSION_COOKIE_DOMAIN: .example.com
+      SAFETY_API_KEY: ${{ secrets.SAFETY_API_KEY }}
+
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
@@ -183,7 +200,7 @@ jobs:
 
       - name: Install dependencies
         working-directory: ./backend
-        run: poetry install --no-root --no-interaction
+        run: poetry sync --no-root --no-interaction
 
       - name: Run Pre-commit hooks
         working-directory: ./backend
@@ -203,6 +220,7 @@ jobs:
           flags: backend
           token: ${{ secrets.CODECOV_TOKEN }}
 
+  # Docker Build/Test
   docker-build:
     name: Docker Build
     runs-on: ubuntu-latest
@@ -230,6 +248,7 @@ jobs:
   security-scan:
     name: Security Analysis
     runs-on: ubuntu-latest
+    environment: portfolio_tests
     needs: [frontend, backend]
     permissions:
       security-events: write
@@ -287,30 +306,10 @@ jobs:
         with:
           python-version: ${{ env.PYTHON_VERSION }}
 
-      - name: Install Poetry
-        uses: snok/install-poetry@v1
+      - name: Run Safety CLI to check for vulnerabilities
+        uses: pyupio/safety-action@v1
         with:
-          version: latest
-          virtualenvs-create: true
-          virtualenvs-in-project: true
-
-      - name: Cache Poetry virtualenv
-        uses: actions/cache@v4
-        with:
-          path: backend/.venv
-          key: ${{ runner.os }}-poetry-${{ hashFiles('backend/poetry.lock') }}
-          restore-keys: |
-            ${{ runner.os }}-poetry-
-
-      - name: Install dependencies
-        working-directory: ./backend
-        run: poetry install --no-root --no-interaction
-
-      - name: Backend safety check
-        working-directory: ./backend
-        run: |
-          poetry run pip install safety
-          poetry run safety check
+          api-key: ${{ secrets.SAFETY_API_KEY }}
 
   image-build-check:
     name: Production Image Build Check
@@ -366,13 +365,14 @@ jobs:
           cache-from: type=gha
           cache-to: type=gha,mode=max
 
+  # PR Status Reporting
   pr-status:
     name: PR Status Report
     runs-on: ubuntu-latest
     needs: [frontend, backend, docker-build, security-scan, image-build-check]
     if: always() && github.event_name == 'pull_request'
     permissions:
-      pull-requests: write
+      issues: write
     steps:
       - name: Report Status
         uses: actions/github-script@v7
