--- backend/.safety-project.ini
@@ -1,4 +1,0 @@
-[project]
-id = backend
-url = /codebases/backend/findings
-name = backend
