+++ backend/pyproject.toml
@@ -19,49 +19,49 @@ security = "scripts:security"
 
 [tool.poetry.dependencies]
 python = ">=3.13,<4.0"
-django = "^6.0.2"
-djangorestframework = "^3.14.0"
-pillow = "^12.1.1"
-django-environ = "^0.13.0"
-django-cors-headers = "^4.0.0"
-psycopg2 = "^2.9.0"
-django-axes = "^8.3.1"
-gunicorn = "^24.1.1"
-django-countries = "^8.2.0"
-django-select2 = "^8.4.8"
-django-ckeditor-5 = "^0.2.20"
-redis = "^5.2.0"
-openai = "^1.0.0"
-django-extensions = "^3.2.0"
-django-parler = "^2.3"
-django-parler-rest = "^2.2"
-beautifulsoup4 = "^4.14.3"
-celery = {extras = ["redis"], version = "^5.4.0"}
-flower = "^2.0.0"
-sentry-sdk = "^2.52.0"
-requests = "^2.32.0"
-django-model-utils = "^5.0.0"
-django-jazzmin = "^3.0.3"
+django = "6.0.2"
+djangorestframework = "3.16.1"
+pillow = "12.1.1"
+django-environ = "0.13.0"
+django-cors-headers = "4.9.0"
+psycopg2 = "2.9.11"
+django-axes = "8.3.1"
+gunicorn = "24.1.1"
+django-countries = "8.2.0"
+django-select2 = "8.4.8"
+django-ckeditor-5 = "0.2.20"
+redis = "5.3.1"
+openai = "1.109.1"
+django-extensions = "3.2.3"
+django-parler = "2.3"
+django-parler-rest = "2.2"
+beautifulsoup4 = "4.14.3"
+celery = {extras = ["redis"], version = "5.6.2"}
+flower = "2.0.1"
+sentry-sdk = "2.52.0"
+requests = "2.32.5"
+django-model-utils = "5.0.0"
+django-jazzmin = "3.0.3"
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^9.0.2"
-pytest-django = "^4.5.0"
-pytest-cov = "^7.0.0"
-coverage = "^7.0.0"
-black = "25.9.0"
+pytest = "9.0.2"
+pytest-django = "4.11.1"
+pytest-cov = "7.0.0"
+coverage = "7.13.3"
+black = "26.3.1"
 isort = "7.0.0"
 flake8 = "7.3.0"
-mypy = "^1.15.0"
-django-stubs = "^5.2.9"
-djangorestframework-stubs = "^3.14.0"
-pre-commit = "^4.3.0"
-safety = "^3.7.0"
-flake8-pyproject = "^1.2.4"
-factory-boy = "^3.3.0"
-Faker = "^37.0.0"
-types-requests = "^2.32.0"
-watchdog = "^6.0.0"
-pytest-mock = "^3.15.1"
+mypy = "1.19.1"
+django-stubs = "5.2.9"
+djangorestframework-stubs = "3.16.8"
+pre-commit = "4.5.1"
+safety = "3.7.0"
+flake8-pyproject = "1.2.4"
+factory-boy = "3.3.3"
+Faker = "37.12.0"
+types-requests = "2.32.4.20260107"
+watchdog = "6.0.0"
+pytest-mock = "3.15.1"
 
 [build-system]
 requires = ["poetry-core"]
