+++ docker/nginx/Dockerfile
@@ -1,4 +1,4 @@
-FROM nginx:latest
+FROM nginx:1.29-alpine
 
 COPY infra/nginx/static_server.conf /etc/nginx/nginx.conf
 COPY frontend/public/error.html /usr/share/nginx/html/error.html
