+++ frontend/src/i18n.server.ts
@@ -9,23 +9,7 @@ import i18next from 'i18next';
 import { initReactI18next } from 'react-i18next';
 import en from '../public/locales/en/translation.json';
 import pl from '../public/locales/pl/translation.json';
-
-type SupportedLanguage = 'en' | 'pl';
-
-function detectLanguage(acceptLanguage: string): SupportedLanguage {
-  if (!acceptLanguage) return 'en';
-
-  const languages = acceptLanguage
-    .split(',')
-    .map(lang => lang.split(';')[0].trim().toLowerCase());
-
-  for (const lang of languages) {
-    if (lang.startsWith('pl')) return 'pl';
-    if (lang.startsWith('en')) return 'en';
-  }
-
-  return 'en';
-}
+import { detectLanguage } from './shared/i18n/detectLanguage';
 
 export async function createServerI18n(acceptLanguage = 'en') {
   const instance = i18next.createInstance();
