+++ infra/nginx/static_server.conf
@@ -27,8 +27,10 @@ http {
     # Rate Limiting Zones
     # - API Zone: 25r/s for modern SPA usage (burst to 50)
     # - Auth Zone: 5r/m for login/reset endpoints
+    # - Contact Zone: 5r/m for the contact form (BFF endpoint)
     limit_req_zone $binary_remote_addr zone=api_limit:10m rate=25r/s;
     limit_req_zone $binary_remote_addr zone=auth_limit:10m rate=5r/m;
+    limit_req_zone $binary_remote_addr zone=contact_limit:10m rate=5r/m;
 
     # Real IP Trust (Trust all standard private network ranges)
     set_real_ip_from 10.0.0.0/8;
@@ -296,6 +298,17 @@ location ^~ /protected_media/ {
         location /health {
             access_log off;
             return 200 'OK';
+        }
+
+        location = /app/contact {
+            limit_req zone=contact_limit burst=2 nodelay;
+            proxy_pass http://frontend_upstream;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Host $host;
+            proxy_set_header X-Forwarded-Proto $proxy_x_forwarded_proto;
+            proxy_http_version 1.1;
         }
 
         location / {
